refactor: moving around contract sync utils
We had contract sync related functionality in a somewhat random place so I moved it to `pxe/src/contract_sync/index.ts`. I also moved the `syncState` function out of `ContractStore` because it doesn't seem to be something that contract store should be handling (stores should be concerned with DB not with utility function simulation).

I decided against introducing `ContractService` because the functionality is stateless. As a result of this wrapping the functionality in a class is somewhat pointless.

I am not super sure about the directory structure and having it all be in `index.ts` so I am open to suggestiom.

This was originally discussed in https://github.com/AztecProtocol/aztec-packages/pull/19840#discussion_r2717251981.
